+++ SRC/FB_MAIN.C
@@ -1,30 +1,44 @@
-#include <stdio.h>
-
+#include "FB_DEFS.H"
+#include "G_PIPE.H"
 #include "G_PLAYER.H"
 #include "S_CONTRL.H"
+#include "S_GSTATE.H"
 #include "S_HBCNTR.H"
-#include "V_VIDEO.H"
-
-static ScreenBuffer_t screen_buffer[2];
+#include "S_MAIN.H"
+#include "V_MAIN.H"
 
 int main(void) {
   Player_t player;
+  Pipe_t pipe;
 
-  // TODO: create inits for all the various subsystems
-  // e.g., video_init, sys_init, etc.
-  v_init_video_subsystem(screen_buffer);
-  s_init_controllers();
+  v_init();
+  s_init();
 
   player = g_create_player(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2, controller_1);
+  pipe = g_create_pipe();
+  add_sprite_to_sprites(&pipe.sprite);
+  add_sprite_to_sprites(&player.sprite);
 
   while (1) {
     s_sys_start();
 
-    g_handle_player_input(&player);
-    g_update_player_sprite_pos(&player);
-    v_display_video(screen_buffer, &player.sprite);
-    FntPrint("dt: %d\n", s_delta_time());
+    switch (s_curr_game_state) {
+      case S_GSTATE_NORMAL:
+        g_player_loop(&player);
 
+        if (player.is_dead) s_curr_game_state = S_GSTATE_GAME_OVER;
+        break;
+      case S_GSTATE_GAME_OVER:
+        FntPrint("Game Over!\n");
+        break;
+      default:
+        FntPrint("Unhandled game state: %d\n", s_curr_game_state);
+        break;
+    }
+
+    v_loop();
+
+    FntPrint("dt: %d\n", s_delta_time());
     s_sys_end();
   }