+++ SRC/G_MAIN.C
@@ -1,0 +1,3 @@
+#include "G_MAIN.H"
+
+void g_init(void) {}