+++ SRC/G_PLAYER.C
@@ -1,12 +1,14 @@
 #include <libetc.h>
 #include <libpad.h>
+#include <stdio.h>
 
 #include "G_PLAYER.H"
 #include "S_HBCNTR.H"
 
-// Approximation for 9.8 m/s / 2.0
-#define GRAVITY 49
+// Approximation for (9.8 / 2.0) m/s
+#define GRAVITY -49
 #define JUMP_VELOCITY 98
+#define VELOCITY_LERP_VAL 7
 
 Player_t g_create_player(short start_x, short start_y, u_char *controller) {
   Player_t p;
@@ -21,22 +23,32 @@ Player_t g_create_player(short start_x, short start_y, u_char *controller) {
 
   p.controller = controller;
 
+  p.velocity = GRAVITY;
+
+  p.is_dead = 0;
+
   return p;
 }
 
 void g_update_player_sprite_pos(Player_t *p) {
-  p->position.vy += GRAVITY * s_delta_time() / 4096;
+  // FIXME: uncomment this later.
+  /*
+  p->position.vy -= p->velocity * s_delta_time () / 4096;
+  if (p->velocity > GRAVITY)
+    p->velocity -= VELOCITY_LERP_VAL;
 
-  // tmp, just put the player back on screen if they're too far gone.
-  if (p->position.vy > 480) p->position.vy = 0;
+  if (p->position.vy > 240 + 8) p->is_dead = 1;
+  */
 
-  // TODO: sprite size shouldn't be hardcoded.
+  // TODO: sprite dimensions shouldn't be hardcoded.
   setXY4(&p->sprite, (p->position.vx) - 8, (p->position.vy) + 8,
          (p->position.vx) + 8, (p->position.vy) + 8, (p->position.vx) - 8,
          (p->position.vy) - 8, (p->position.vx) + 8, (p->position.vy) - 8);
 }
 
+#define JUMP_BUTTON_MASK PADRdown
 void g_handle_player_input(Player_t *p) {
+  static int last_padd;
   int padd;
 
   if (p->controller[0] != 0) {
@@ -46,5 +58,13 @@ void g_handle_player_input(Player_t *p) {
 
   padd = ~((p->controller[2] << 8) | (p->controller[3]));
 
-  if (padd & PADRdown) p->position.vy -= JUMP_VELOCITY * s_delta_time() / 4096;
+  if ((padd & JUMP_BUTTON_MASK) && !(last_padd & JUMP_BUTTON_MASK))
+    p->velocity = JUMP_VELOCITY;
+
+  last_padd = padd;
+}
+
+void g_player_loop(Player_t *p) {
+  g_handle_player_input(p);
+  g_update_player_sprite_pos(p);
 }