+++ SRC/S_MAIN.C
@@ -1,0 +1,4 @@
+#include "S_CONTRL.H"
+#include "S_MAIN.H"
+
+void s_init(void) { s_init_controllers(); }